Moved Firebase Collection Query to SearchScreen
Moved the function responsible for retrieving the service providers' collection from the ProvidersInfoList component to the SearchScreen. This modification was made to enable the SearchBar component to access the collection and perform provider searches efficiently.
