+++ CtasApp/src/components/ProviderInfoList.tsx
@@ -1,37 +1,11 @@
-import React, {useState, useEffect} from 'react';
+import React from 'react';
 import {FlatList, View} from 'react-native';
 import {SearchStyles} from '../styles/SearchStyles';
 import ProviderInfoCard from '../components/ProviderInfoCard';
-import {useUserContext} from '../../UserContext';
-import firestore from '@react-native-firebase/firestore';
-import {Provider} from '../schema/ProviderSchema';
 import {truncateString} from '../helpers/TruncateStringHelper';
-
-type ProviderWithKey = Provider & {key: string};
-
-const ProviderInfoList: React.FC = () => {
-  const [providers, setProviders] = useState<ProviderWithKey[]>([]);
-  const {userType} = useUserContext();
+import { ProvidersListProps } from '../schema/ProvidersListSchema';
+const ProviderInfoList: React.FC<ProvidersListProps>  = ({ providers }) => {
   const cardProviderSeparator = () => <View style={SearchStyles.separator} />;
-
-  useEffect(() => {
-    const unsubscribe = firestore()
-      .collection('Providers')
-      .onSnapshot(querySnapshot => {
-        const providersList: ProviderWithKey[] = [];
-        querySnapshot.forEach(documentSnapshot => {
-          const provider = documentSnapshot.data() as Provider;
-          providersList.push({
-            ...provider,
-            key: documentSnapshot.id,
-          });
-        });
-        setProviders(providersList);
-      });
-
-    return () => unsubscribe();
-  }, []);
-
   return (
     <FlatList
       data={providers}
