+++ CtasApp/src/schema/ProvidersListSchema.ts
@@ -1,0 +1,5 @@
+import {Provider} from '../schema/ProviderSchema';
+type ProviderWithKey = Provider & {key: string};
+export interface ProvidersListProps {
+  providers: ProviderWithKey[],
+}
